requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
rapidfuzz>=3.5.0
holidays>=0.50
